issue #544: Refactor storage volume download
The current storage volume download implementation relies on writing the
new volume as a file. But that's not the case for some storage pool
types (e.g. logical), which don't treat their volumes as plain files.

Use the libvirt API to write data to a storage volume instead of writing
it directly to a file. The libvirt API requires an existing volume to
write the data to it, and a volume requires a capacity to be created.
As some remote URLs don't provide the header 'Content-length', sometimes
we need to download a file without knowing its size in advance. Due to
those cases, instead of downloading and writing the data directly to the
storage volume, this patch makes it so a temporary file is always
created - so we can get its size - and then that file is transfered to a
storage volume.

Fix issue #544 ("Download to logical storage pool fails").

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>