Add a disk to the VM when testing snapshot
The test case which tests snapshots creates a VM without any disk and
then creates a snapshot of it. However, newer versions of libvirt don't
allow snapshots to be taken of shutoff domains which don't contain any
disk - because there's nothing to take snapshot of.

Update the test case "ModelTests.test_vm_lifecycle" so the VM has one
disk and a snapshot of it can be created.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>