MockModel: Extend logical storage pool
Model does not use libvirt to extend the logical pool. Instead of that,
the LVM is extended by using vgextend command.
As in MockModel environment we don't have a LVM itself, we need to fake
the extend operation. In that case only the pool XML is updated without
affecting the current system.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>