Remove workaround when deleting a VM due to tests
Old versions of tests used libvirt's Test driver directly and some
functions didn't work with it. In those cases, we needed a workaround so
those functions were able to work with the real model and the test model.
However, that's not the case anymore because we overwrite the functions
which don't make sense in a test environment. so some workarounds aren't
needed anymore.

When deleting a VM, do not handle an exception which we only expect when
running Kimchi on libvirt's Test driver. The current code uses an
overwritten, custom function which makes the workaround obsolete.

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>