Handle requests differently for transient objects
If a transient (or non-persistent) object (e.g. VM, network or
storage pool) is created outside of Kimchi, it will (expectedly)
be deleted when shutdown. Since Kimchi calls a GET after
shutdown and other POST requests, the GET fails, and a user will
receive an error. However, this is not an error condition.

This patch does not call the GET function for transient objects
if the function will delete the object (as indicated by a new
parameter passed to generate_action_handler).

Signed-off-by: Christy Perez <christy@linux.vnet.ibm.com>