issue #595: Return correct memory value when VM is shutoff
Newer libvirt versions report 0 when reading the current memory of a
shutoff VM by calling the method "domain.info()". So when the VM is
shutoff, the REST command "GET /vms/<vm>" is returning an invalid
memory value.

Read the memory value from the VM's XML descriptor when it is shutoff,
and keep returning the value from "domain.info()" otherwise.

Fix issue #595 (Memory is set to 0 (null) when editing guest).

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>