bug fix: Allow adding a iSCSI/SCSI volume from a non-ASCII pool to a template
When setting a iSCSI/SCSI pool to a template, the user must also specify
the volume as those pools are read-only.
But it was not possible when the targeted pool name had non-ASCII
characteres.
That is because the libvirt API is based on string and Kimchi on
unicode. So when using Kimchi internal functions we must use unicode.
To fix it convert the string returned by libvrit API to unicode.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>