issue #564: Parse logical volumes to find out their actual formats
All storage volumes belonging to a logical storage pool are reported by
libvirt as 'raw', even when they are ISO images. When creating a
template, only the volumes listed as 'iso' are displayed, so it's currently
not possible to create a template using a volume from a logical storage
pool.

Instead of relying only on the volume format reported by libvirt, open
and inspect the storage volumes which belong to a logical storage pool
in order to find out whether they're ISO images.

Fix issue #564 (List volumes other than "format=iso" when creating
a template).

Signed-off-by: Crístian Viana <vianac@linux.vnet.ibm.com>