Create empty files on rpm build
Issue #496 - System leftovers after uninstalling Kimchi

A rpm does not remove files that are not created by the installation. So,
all the files were provided by the package and now they can be
uninstalled. Beyond this, the function that creates the proxy conf
now accepts empty files to fill the configuration.

Signed-off-by: Ramon Medeiros <ramonn@linux.vnet.ibm.com>