Use default network model when attaching a NIC
When attaching a network interface to a virtual machine, Kimchi lets
libvirt choose the network model. However, there is an existing lookup
table which can determine the best network model for each type of guest
operating system, and Kimchi should use that information here as well.

Use the osinfo lookup table to determine the best possible network
interface model when attaching a NIC to a VM.

Signed-off-by: Crístian Deives <cristiandeives@gmail.com>