Template: Assign 'raw' to disk format if pool is 'logical' or [i]scsi
Kimchi always creates raw disk format to new vm disks if the storagepool
is logical, iscsi or scsi. But, new templates where not using this
logic and was possible, through API calls, to create a template with a
SCSI storagepool and disk format vmdk, for instance.
This patch fixes this issue and makes necessary change in one of the
tests in order to avoid an error with _get_storage_type() and
deactivated storagepools.

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>