Replace storage volume 'ref_cnt' with 'used_by'
Instead of storing the number of VMs which use each storage volume
('ref_cnt'), we should store the names of the VMs which use each volume
('used_by'). That new variable can provide the same information to the user
while also providing more details.

Use the variable 'used_by' to store the names of the VMs which use a
storage volume instead of 'ref_cnt' which stores the number of VMs which
use a volume.

Signed-off-by: Crístian Deives <cristiandeives@gmail.com>