Remove unecessary call to node_dev.parent()
On hostdev.py, device parent was being search 2 times. First, a xml was
parsed, where the parent device can be found. When device parent not
present, calling the function .parent() (which is not necessary) will raise
an error on libvirt. To avoid this, the call for .parent() was removed,
and parent device is set as None if not present.

Signed-off-by: Ramon Medeiros <ramonn@linux.vnet.ibm.com>