Set default VM template memory to 2048 in Power
There are no official Kimchi support for VMs running with under 2048Mb
of RAM in Power systems. This patch set the default memory of templates
created in Power hosts to 2048, instead of 1280.

The 'memory' values were removed from the template_specs dict because
it is no longer supported - the default value will be either '1024' for
x86, '2048' for Power or any value defined in template.conf.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>