Raw volumes validation: back-end and front-end
'Raw' volumes can not be valid image disks (e.g. XML, PDF, TXT are
all raw files), so it's necessary check the 'content' of them and
return the information to GET requests.

Updated front-end to make a better list of volume options on VM's
edition screen based on GET response of storage volumes.

Updated MockModel to reflect the changes and use it correctly on tests.

This patch solves the issue #731.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>