Issue #740: Solving UI dependencies checking.
Commit 9fbb105 added pip commands on Makefile to install some Python modules
during the UI compilation. However, the compilation step of the UI files is
executed before the block where the commands were added, resulting in
compilation fail if the system doesn't have the modules installed. In addition,
pip commands need to be executed as root and, as good practice, the compilation
should be executed as normal user.

Removing from Makefile the pip commands to instal the Python modules and adding
checks in configure.ac to see if they are installed. Also, edit the README.md
to explain how to install the modules as dependencies to build Wok.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>