Fix issue 766 - Define network interface in libvirt
When a network interface was only visible to the system and
not to libvirt, libvirt throws an error when user asks a
bridged network to be created using that interface, since
it is not visible to libvirt.

This patch adds network interface redefinition code to make
it visible to libvirt, allowing the bridged network to be
created.

Signed-off-by: Lucio Correia <luciojhc@linux.vnet.ibm.com>