Add validation and fix parameters in _generate_action_handler_base
The function _generate_action_handler_base is used in
asynchronous tasks like clone and migration. It usually
handles POST requests, which may contain parameters.

Handler was not able to differentiate the parameters, so
user was forced to pass all and in the right position,
also, parameters where not checked against API.json.

This patch do the following changes to fix the issues:
- handle parameters, and if not passed, sends None
- check parameters against API.json (adds validate_params)

Signed-off-by: Rodrigo Trujillo <rodrigo.trujillo@linux.vnet.ibm.com>