Bug fix: Respect user request while building the UI
Every time the UI was built (for a first access or after a browser
refresh), the user was always redirected to the first tab.
Fix it, by respecting the user request when building the UI and only
redirect to the default page when no specific tab was requested.

This patch also fixes issue #3

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>